requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.29
python-dotenv==1.0.0
httpx==0.25.2
APScheduler==3.10.4
